refactor(sdk,scopes): move with* utils to toolkit folder
